+++ src/main/java/org/folio/linked/data/repo/ProfileRepository.java
@@ -4,6 +4,6 @@ import java.util.List;
 import org.folio.linked.data.model.entity.Profile;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProfileRepository extends CrudRepository<Profile, Long> {
+public interface ProfileRepository extends CrudRepository<Profile, Integer> {
   List<Profile> findByResourceTypeUri(String resourceTypeUri);
 }
